vsr: quicker request protocol
- Client broadcasts the first .register request to quickly learn a view.
  - _retries_ of .register are not broadcast, to reduce cluster load.
- When retrying, simultaneously:
  - re-send the request to the primary, to account for potential message
    loss
  - send a ping to a single replica in a round-robin fashion, to account
    for a potential view change

Consequently, backups no longer try to forward request messages. We
maintain the invariant  that the client always talks directly to the
primary and don't amplify the network.

I also considered resetting the request timeout when the client jumps
view, but figured that it's better to let the client ride the timeout
normally, to avoid potential thundering herd after view change.